+++ src/index.ts
@@ -9,15 +9,17 @@ export {
   EthereumCallContractPayload,
   EthereumDeployContractPayload,
   Peer,
-  SwapRequest
+  SwapRequest,
+  SwapDetails
 } from "./cnd/cnd";
 export * from "./cnd/siren";
 
 export { Actor, createActor } from "./actor";
 
 export { BitcoinWallet, InMemoryBitcoinWallet } from "./bitcoin_wallet";
-
 export { EthereumWallet } from "./ethereum_wallet";
+export { LightningWallet, Outpoint } from "./lightning_wallet";
+
 export { BigNumber } from "bignumber.js";
 
 export { ComitClient } from "./comit_client";
@@ -35,3 +37,4 @@ export { MakerNegotiator } from "./negotiation/maker/negotiator";
 export { TakerNegotiator } from "./negotiation/taker/negotiator";
 
 export { TryParams } from "./util/timeout_promise";
+export { Lnd } from "./lnd";