+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add lightning support for swap actions.
+
 ## [0.12.0] - 2020-02-25
 
 ### Added